Rename 'value' to 'content'
Why? I argue the word content can be understood more easily than value. The word "value" can be attached to many things e.g. also to field names. The word "content" on the other hand means a value is within something and therefore in this specific context would be more specific.
