+++ src/main/resources/l10n/JabRef_en.properties
@@ -2496,20 +2496,20 @@ Success\!\ Finished\ writing\ metadata.=Success! Finished writing metadata.
 Custom\ API\ key=Custom API key
 Check\ %0\ API\ Key\ Setting=Check %0 API Key Setting
 
-Edit\ field\ value=Edit field value
+Edit\ field\ content=Edit field content
 Two\ fields=Two fields
 Overwrite\ Non\ empty\ fields=Overwrite Non empty fields
 Set=Set
 Append=Append
-Clear\ field=Clear field
-Field\ value=Field value
-Edit\ field\ value\ of\ selected\ entries=Edit field value of selected entries
+Clear\ field\ content=Clear field content
+Set\ or\ Append\ content=Set or Append content
+Edit\ field\ content\ for\ selected\ entries=Edit field content for selected entries
 Rename=Rename
 New\ field\ name=New field name
-Copy\ value=Copy value
-Move\ value=Move value
-Swap\ values=Swap values
-Copy\ or\ move\ the\ value\ of\ one\ field\ to\ another=Copy or move the value of one field to another
+Copy\ content=Copy content
+Move\ content=Move content
+Swap\ content=Swap content
+Copy\ or\ move\ the\ content\ of\ one\ field\ to\ another=Copy or move the content of one field to another
 Automatic\ field\ editor=Automatic field editor
 
 (Note\:\ If\ original\ entries\ lack\ keywords\ to\ qualify\ for\ the\ new\ group\ configuration,\ confirming\ here\ will\ add\ them)=(Note: If original entries lack keywords to qualify for the new group configuration, confirming here will add them)
